[POA-507] Prepare TS package for NPM publishing (#11)
This PR prepares the Typescript code to be released as a private NPM
package that will be accessible by the Postman NPM org. This should help
us improve the frontend's mock and test utilities initially proposed in
https://github.com/postman-eng/postman-test-data-manager/pull/65 by
allowing us to build and serialize mock witnesses for different E2E test
cases.